added config module
